Major refactor applying SW engineering principles
Removed all the repeated code in the patient_journey method by extracting the wait for resource and the wait for service time into a separate method

Removed a bug in Patient class by setting the ACU/ED choice in the init. As it was written, using the method, determine_acu_destiny - there’s a high chance (~30%) of getting a different value, which is obviously incorrect.

Also corrected the line: 
if patient.acu_patient == True

acu_patient is already a boolean, all that’s required is:

if patient.acu_patient:

I renamed the attribute to make this clearer, so it’s now:

if patient.is_acu_patient:

Re-wrote the statistics collection. Appending pandas dataframes is a no-no because it involves massive data copying. Not a problem here with a small data set, but not a practice to be encouraged. Instead keep a simple python list holding dictionaries of the different service wait times for every patient. Once the trial is over, convert this to a pandas data frame and write to CSV

I changed the calculation from mean to 90th percentile, which is more pertinent when looking at queues.

Used python to automate initialisation, notably with comprehensions

Used f-strings to tidy up the reports